% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarizedData.R
\docType{methods}
\name{as.background}
\alias{as.background}
\alias{as.background,summarizedData-method}
\title{generate background data}
\usage{
as.background(object)

\S4method{as.background}{summarizedData}(object)
}
\arguments{
\item{object}{summarizedData-object}
}
\value{
list with background data for the filter plots
}
\description{
Generate background data from a summarizedData-object
}
\details{
Generates a background dataset can be used in the filter plots
}
\author{
mvaniterson
}

