% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarizedData.R
\docType{methods}
\name{combine,summarizedData,ANY-method}
\alias{combine,summarizedData,ANY-method}
\title{concatenate two summarizedData objects into one object}
\usage{
\S4method{combine}{summarizedData,ANY}(x, ..., by = c("identical", "overlap"))
}
\arguments{
\item{x}{summarizedData-object}

\item{...}{onne or more summarizedData-object}

\item{by}{argument indicating how the targets information should be combined}
}
\value{
one summarizedData object
}
\description{
concatenate two summarizedData objects into one object
}
\examples{
data(exampleData)
combine(exampleData, exampleData)
}

