% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarizedData.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,summarizedData-method}
\title{concatenate two or more summarizedData objects into one object}
\usage{
combine(object, ..., by = c("identical", "overlap"))

\S4method{combine}{summarizedData}(object, ..., by = c("identical",
  "overlap"))
}
\arguments{
\item{object}{summarizedData-object}

\item{...}{one or more summarizedData-objects}

\item{by}{argument indicating how the targets information should be combined}
}
\value{
one summarizedData object
}
\description{
combine summarizedData objects
}
\examples{
data(exampleData)
combine(exampleData, exampleData)
}

