\name{exampleData}
\docType{data}
\alias{exampleData}
\title{summarizedData object on 500 450k Human Methylation samples}
\description{Pre-summarizedData object on 500 450k Human Methylation
  samples. Can be used as input for visualize}
\usage{exampleData}
\value{Pre-summarizedData object on 500 450k Human Methylation
  samples.}
\format{summarizedData-object}
\keyword{datasets}
\examples{
data(exampleData)
\dontrun{visualize(exampleData)} 
}
  
