% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizedData.R
\docType{methods}
\name{show,summarizedData-method}
\alias{show,summarizedData-method}
\title{show method for Illumina Human DNA Methylation array data}
\usage{
\S4method{show}{summarizedData}(object)
}
\arguments{
\item{object}{summarizedData object}
}
\value{
print short summary summarizedData object
}
\description{
show method for summarized Illumina Human DNA Methylation array data
}
\examples{
data(exampleData)
exampleData
}
