% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizedData.R
\docType{methods}
\name{combine,summarizedData,ANY-method}
\alias{combine,summarizedData,ANY-method}
\title{concatenates two summarizedData objects into one object}
\usage{
\S4method{combine}{summarizedData,ANY}(x, y, by = c("identical", "overlap"))
}
\arguments{
\item{x}{summarizedData-object}

\item{y}{summarizedData-object}

\item{by}{argument indicating how the targets information should be combined}
}
\value{
one summarizedData object
}
\description{
Concatenates two summarizedData objects into one object
}
\examples{
data(exampleData)
combine(exampleData, exampleData)
}
