% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summarizedData.R
\docType{methods}
\name{show,summarizedData-method}
\alias{show,summarizedData-method}
\title{show method for summarized 450k Illumina Human Methylation data}
\usage{
\S4method{show}{summarizedData}(object)
}
\arguments{
\item{object}{summarizedData object}
}
\value{
print short summary summarizedData object
}
\description{
show method for summarized 450k Illumina Human Methylation data
}

