% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{visualize}
\alias{visualize}
\alias{visualize,summarizedData-method}
\title{visualize the summarized 450k data}
\usage{
visualize(object, thresholds = list(MU = 10.5, OP = 11.75, BS = 12.75, HC =
  13.25, DP = 0.95), background = NULL)

\S4method{visualize}{summarizedData}(object, thresholds = list(MU = 10.5, OP =
  11.75, BS = 12.75, HC = 13.25, DP = 0.95), background = NULL)
}
\arguments{
\item{object}{summarizedData object}

\item{thresholds}{default thresholds}

\item{background}{optional summarizedData-object used as background in filter control plots}
}
\value{
lauches a web browser with the shiny application and returns a
data.frame with detected outliers
}
\description{
launch a shiny app for visualization of the summarized 450k data
}
\details{
Outliers are detected based on a set of default thresholds.
To use a use-defined set of thresholds use the thresholds
argument.
}
\examples{
library(minfiData)
baseDir <- system.file("extdata", package="minfiData")
targets <- read.450k.sheet(baseDir)
data <- summarize(targets)
\dontrun{
visualize(data)
}
}

